Locations UI updates
TO DO:

Change colours of app when destroyed
Clean code
Move locations to top bar

Optional
Runnable last updated
Themes
