+++ app/src/main/res/layout/fragment_locations_map.xml
@@ -5,30 +5,20 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.location.LocationsMapFragment">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/googleMap"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:name="com.google.android.gms.maps.SupportMapFragment"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
+    tools:context=".ui.location.LocationsMapFragment"
+    android:background="@color/background_white">
 
     <androidx.cardview.widget.CardView
-        android:layout_width="0dp"
-        android:layout_height="40dp"
-        app:cardCornerRadius="30dp"
+        android:id="@+id/searchBar"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        app:cardCornerRadius="5dp"
         android:layout_marginTop="16dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginEnd="20dp"
+        android:layout_marginHorizontal="16dp"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintWidth_percent="0.7"
-        app:cardBackgroundColor="@color/search_bar"
+        app:cardBackgroundColor="@color/card_white"
         >
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/autocomplete_fragment"
@@ -37,19 +27,36 @@
             android:name="com.google.android.libraries.places.widget.AutocompleteSupportFragment"/>
     </androidx.cardview.widget.CardView>
 
-    <Button
-        android:id="@+id/btnSetLocation"
-        android:layout_width="wrap_content"
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:textSize="16sp"
-        android:text="Set Location"
-        android:visibility="invisible"
-        android:backgroundTint="@color/gradient_start"
+        app:layout_constraintTop_toBottomOf="@id/searchBar">
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/googleMap"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
+        app:layout_constraintTop_toBottomOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/btnSetLocation"
+        android:layout_width="match_parent"
+        android:layout_height="70dp"
+        android:fontFamily="@font/fira_sans_medium"
+        android:textSize="16sp"
+        android:text="Add Location"
+        android:visibility="gone"
+        android:backgroundTint="@color/blue_grey"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.926"/>
+        android:layout_marginHorizontal="16dp"
+        android:layout_marginBottom="16dp"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
