+++ app/src/main/res/layout/location_item.xml
@@ -1,103 +1,45 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView
+<androidx.cardview.widget.CardView
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/tempCard"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    style="?attr/materialCardViewFilledStyle"
+    android:layout_height="100dp"
+    android:layout_margin="16dp"
+    app:cardBackgroundColor="@color/card_white"
+    android:clickable="true"
+    android:foreground="@drawable/selectable_background_ripple"
     app:cardCornerRadius="16dp"
-    tools:ignore="UnusedAttribute">
+    app:cardElevation="0dp"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent">
 
-    <androidx.cardview.widget.CardView
-        android:id="@+id/container"
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:foreground="@drawable/selectable_background_ripple"
-        tools:ignore="UnusedAttribute">
+        android:layout_height="match_parent"
+        android:padding="18dp"
+        android:orientation="vertical">
 
-        <LinearLayout
-            android:id="@+id/item"
-            android:layout_width="match_parent"
+        <TextView
+            android:id="@+id/tvArea"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:background="?attr/colorSurface"
-            android:orientation="horizontal"
-            android:gravity="center_vertical">
+            android:fontFamily="@font/fira_sans_medium"
+            android:text="Area"
+            android:textColor="@color/black"
+            android:textSize="20sp"
+            android:layout_gravity="center_vertical"/>
 
-            <LinearLayout
-                android:id="@+id/content"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
+        <TextView
+            android:id="@+id/tvCountry"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Country"
+            android:textColor="@color/dark_grey"
+            android:textSize="15sp"
+            android:layout_gravity="center_vertical"/>
 
-                <RelativeLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="16dp"
-                    android:layout_marginBottom="2dp"
-                    android:layout_marginEnd="16dp"
-                    tools:ignore="UseCompoundDrawables">
+    </LinearLayout>
 
-                    <ImageView
-                        android:id="@+id/weather_icon"
-                        android:layout_width="40dp"
-                        android:layout_height="40dp"
-                        android:layout_marginEnd="16dp"
-                        android:layout_centerVertical="true"
-                        android:src="@drawable/sun"
-                        tools:ignore="ContentDescription" />
-
-                    <androidx.appcompat.widget.LinearLayoutCompat
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_centerVertical="true"
-                        android:layout_toEndOf="@id/weather_icon"
-                        android:orientation="vertical">
-
-                        <TextView
-                            android:id="@+id/tvArea"
-                            android:layout_width="match_parent"
-                            android:layout_height="wrap_content"
-                            android:maxLines="1"
-                            android:ellipsize="end"
-                            android:textAlignment="viewStart"
-                            android:textColor="@color/black"
-                            tools:text="Dubai"
-                            tools:ignore="RtlHardcoded" />
-
-                        <TextView
-                            android:id="@+id/tvCountry"
-                            android:layout_width="match_parent"
-                            android:layout_height="wrap_content"
-                            android:maxLines="1"
-                            android:ellipsize="end"
-                            android:textAlignment="viewStart"
-                            android:textColor="@color/black"
-                            tools:text="United Arab Emirates"
-                            tools:ignore="RtlHardcoded" />
-
-                    </androidx.appcompat.widget.LinearLayoutCompat>
-
-                </RelativeLayout>
-
-                <TextView
-                    android:id="@+id/source"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginEnd="16dp"
-                    android:layout_marginBottom="16dp"
-                    android:textStyle="bold"
-                    android:textSize="10sp"
-                    android:maxLines="1"
-                    android:ellipsize="end"
-                    android:textAlignment="viewStart"
-                    tools:text="Data by Open-Meteo"
-                    tools:ignore="RtlHardcoded,SmallSp" />
-
-            </LinearLayout>
-
-        </LinearLayout>
-
-    </androidx.cardview.widget.CardView>
-
-</com.google.android.material.card.MaterialCardView>
+</androidx.cardview.widget.CardView>
